[BENCH-826] Move MockAwsApi, MockAzureApi, MockGcpApi to mocks package
