+++ service/src/test/java/bio/terra/workspace/common/mocks/MockAzureApi.java
@@ -1,4 +1,4 @@
-package bio.terra.workspace.common.utils;
+package bio.terra.workspace.common.mocks;
 
 import org.springframework.stereotype.Component;
